refactor: fsk utils calculates a whole number int

+++ Core/Inc/utils/fsk_utils.h
@@ -1,6 +1,6 @@
 #ifndef FSK_UTILS_H
 #define FSK_UTILS_H
 
-double calculate_sample_rate(double f1, double f2, double baud);
+int calculate_sample_rate(double f1, double f2, double baud);
 
 #endif // FSK_UTILS_H
